fixed us unnessesarily truncating our id fields

+++ server/model/crontab.go
@@ -11,7 +11,7 @@ func (Crontab) TableName() string {
 
 // Crontab struct is a row record of the crontab table in the tca database
 type Crontab struct {
-	Cron     int32       `gorm:"primary_key;AUTO_INCREMENT;column:cron;type:int;" json:"cron"`
+	Cron     int64       `gorm:"primary_key;AUTO_INCREMENT;column:cron;type:int;" json:"cron"`
 	Interval int32       `gorm:"column:interval;type:int;default:7200;" json:"interval"`
 	LastRun  int32       `gorm:"column:lastRun;type:int;default:0;" json:"last_run"`
 	Type     null.String `gorm:"column:type;type:enum ('news', 'mensa', 'kino', 'roomfinder', 'alarm', 'fileDownload','dishNameDownload','averageRatingComputation', 'iosNotifications', 'iosActivityReset', 'canteenHeadCount');" json:"type"`
